feat(constants): rename enums file to constants
