+++ src/facturapi/__init__.py
@@ -1,3 +1,1 @@
 from .wrapper import Facturapi
-
-__all__ = ["Facturapi"]
